Rename marshalling.py -> error_store.py
